requirements.txt
asyncio==3.4.3
async_timeout==3.0.1
uvloop==0.11.2

Django==2.1.2.
mysqlclient==1.3.13

librouteros==2.1.1

python-daemon==2.2.0
pyTelegramBotAPI==3.6.6



# -------
attrs==18.2.0
voluptuous==0.11.5
